Add files via upload
Project 1 of the WorldQuandant university projects.
Dataset: A Mexico Housing CSV for real Estate.
After wrangling and cleaning the tabular data, I performed some Exploratory Analysis on it to discover  the distribution of the houses, states or regions that are more prevalent, where Mexico's biggest cities are based on this distribution.
Moving on 
RESEARCH QUESTIONS.
1)Which state has the most expensive real estate market?
Do housing prices vary by state? If so, which are the most expensive states for purchasing a home? During our exploratory data analysis, we used descriptive statistics like mean and median to get an idea of the "typical" house price in Mexico. Now, we need to break that calculation down by state and visualize the results.
2)s there a relationship between home size and price?

From our previous question, we know that the location of a home affects its price (especially if it's in Mexico City), but what about home size? Does the size of a house influence price?

Using a scatter plot we were  able to identify the above Questions.